+++ src/services/users/globalRoles.ts
@@ -1,3 +1,4 @@
+import { rqlBuilder } from '../../rql';
 import type { HttpInstance } from '../../types';
 import { HttpClient } from '../http-client';
 import type { UsersGlobalRolesService } from './types';
@@ -10,10 +11,29 @@ export default (
     return (await client.get(httpWithAuth, '/permissions', options)).data;
   },
 
-  async get(options) {
+  async find(options) {
     return (
       await client.get(httpWithAuth, `/roles${options?.rql || ''}`, options)
     ).data;
+  },
+
+  async findFirst(options) {
+    const res = await this.find(options);
+    return res.data[0];
+  },
+
+  async findById(id, options) {
+    const rqlWithId = rqlBuilder(options?.rql).eq('id', id).build();
+    return await this.findFirst({ ...options, rql: rqlWithId });
+  },
+
+  async findByName(name, options?) {
+    const rqlWithName = rqlBuilder(options?.rql).eq('name', name).build();
+    return await this.findFirst({ ...options, rql: rqlWithName });
+  },
+
+  async get(options) {
+    return this.find(options);
   },
 
   async create(requestBody, options) {